pyproject.toml
[project]
name = "openmindat-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix-otel>=0.13.1",
    "fastmcp>=2.11.3",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.6",
    "langgraph-supervisor>=0.0.29",
    "openinference-instrumentation-langchain>=0.1.52",
    "openinference-instrumentation-mcp>=1.3.1",
    "openmindat>=0.1.2",
]
